Refactor ECS core, improve thread safety, add tests
- Switch Scene IDs to ushort, update invalid value logic
- Add thread safety to SceneManager and ComponentRegistry
- Add GHOST_ZERO_INIT_COMPONENT define for editor configs
- Update mesh header counts to use int, not uint
- Add Collect methods for archetype/component cleanup
- Add With/Without/AsView to ComponentSet for easier use
- Refactor World creation/destruction, version handling
- Refactor ResourceStreamingContext to use init-only props
- Add unit tests for entity queries and world lifecycle
- Minor code cleanups and formatting fixes

+++ src/Editor/Ghost.Editor.Core/Assets/MeshAssetHandler.cs
@@ -2,6 +2,7 @@ using Ghost.Core;
 using Ghost.Core.Utilities;
 using Ghost.Editor.Core.Services;
 using Ghost.Engine;
+using Ghost.Engine.Streaming;
 using Ghost.Graphics.Core;
 using Ghost.Graphics.RHI;
 using Misaki.HighPerformance.Jobs;
@@ -384,16 +385,16 @@ internal class MeshAssetHandler : IImportableAssetHandler, IPackableAssetHandler
         {
             magic = MeshContentHeader.MAGIC,
             version = MeshContentHeader.VERSION,
-            vertexCount = (uint)geometry.Vertices.Count,
-            indexCount = (uint)geometry.Indices.Count,
-            materialPartCount = (uint)geometry.MaterialParts.Length,
-            meshletCount = (uint)meshletData.GetRef().meshlets.Count,
-            meshletGroupCount = (uint)meshletData.GetRef().groups.Count,
-            meshletHierarchyNodeCount = (uint)meshletData.GetRef().hierarchyNodes.Count,
-            meshletVertexCount = (uint)meshletData.GetRef().meshletVertices.Count,
-            meshletTriangleCount = (uint)meshletData.GetRef().meshletTriangles.Count,
-            materialSlotCount = (uint)meshletData.GetRef().materialSlotCount,
-            lodLevelCount = (uint)meshletData.GetRef().lodLevelCount,
+            vertexCount = geometry.Vertices.Count,
+            indexCount = geometry.Indices.Count,
+            materialPartCount = geometry.MaterialParts.Length,
+            meshletCount = meshletData.GetRef().meshlets.Count,
+            meshletGroupCount = meshletData.GetRef().groups.Count,
+            meshletHierarchyNodeCount = meshletData.GetRef().hierarchyNodes.Count,
+            meshletVertexCount = meshletData.GetRef().meshletVertices.Count,
+            meshletTriangleCount = meshletData.GetRef().meshletTriangles.Count,
+            materialSlotCount = meshletData.GetRef().materialSlotCount,
+            lodLevelCount = meshletData.GetRef().lodLevelCount,
             boundsMin = bounds.Min,
             boundsMax = bounds.Max,
         };
